feat: convert QuarkusTemplate to interface and add TemplateRegistry.
